Support routing graphs beyond 32-bit edge IDs

+++ engines/vtr/libs/librrgraph/src/base/rr_graph_fwd.h
@@ -16,7 +16,11 @@ class t_rr_graph_storage;
 class RRGraph;
 
 typedef vtr::StrongId<struct rr_node_id_tag, uint32_t> RRNodeId;
-typedef vtr::StrongId<struct rr_edge_id_tag, uint32_t> RREdgeId;
+// Large, IO-limited emulation partitions can exceed 2^32 routing-resource
+// edges even though their node count remains within 32 bits.  Keep edge IDs
+// wide enough for the complete graph; otherwise partition_edges() silently
+// truncates its StrongId range and aborts while sorting the edge arrays.
+typedef vtr::StrongId<struct rr_edge_id_tag, uint64_t> RREdgeId;
 typedef vtr::StrongId<struct rr_indexed_data_id_tag, uint32_t> RRIndexedDataId;
 typedef vtr::StrongId<struct rr_switch_id_tag, uint16_t> RRSwitchId;
 typedef vtr::StrongId<struct rr_segment_id_tag, uint16_t> RRSegmentId;
